% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fixef.R
\name{create_fixef}
\alias{create_fixef}
\title{Extract the fixed effects for each market}
\usage{
create_fixef(model)
}
\arguments{
\item{model}{}
}
\description{
Extract the fixed effects for each market
}

