% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_glance_eq.R
\name{create_glance_eq}
\alias{create_glance_eq}
\title{Extract some info about the equations}
\usage{
create_glance_eq(model)
}
\arguments{
\item{model}{}
}
\description{
Extract some info about the equations
}

