% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_predictions.R
\name{create_predictions}
\alias{create_predictions}
\title{Create predictions using make_predictions}
\usage{
create_predictions(working_df, df_use, area_todo)
}
\arguments{
\item{area_todo}{}
}
\description{
This takes the coefficients in the working\\_df, joins on the actual data we
had set up. Then it calculates the predictions using the make predictions
function, and iterating down each row of working\\_df.
}
\examples{


}

