% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_plm.R
\name{estimate_model_plm}
\alias{estimate_model_plm}
\title{Estimate model, plm}
\usage{
estimate_model_plm(working_df, df_use)
}
\arguments{
\item{working_df}{a dataframe that contains the equations specs, etc. It
needs to have a column with train_end_date that will used to filter the data
as well as a spec column that can be used to generate the equation specification}

\item{df_use}{a dataframe with the historical data on the markets. It needs to have an
area_sh column and a data column to be used in creating the panel dataframe}
}
\value{
Original dataframe, with addition of a train_panel column and a model column.
}
\description{
Similar to eviews with "Fixed effects" for the intercept. So this uses separate
fixed effects estimate of intercept for each metro
}
\details{
The result will be to add
}
\examples{



}

