% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coef_by_mkt.R
\name{extract_coef_by_mkt}
\alias{extract_coef_by_mkt}
\title{Extract coefficients by market from model}
\usage{
extract_coef_by_mkt(df_with_models)
}
\arguments{
\item{df_with_models}{}
}
\description{
Extract coefficients by market from model
}
\examples{
This isnt necessary to run predictions. Instead it provides a way to access
the coefficients by market, if one wanted to.
}

