% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model_plm_stats.R
\name{extract_model_plm_stats}
\alias{extract_model_plm_stats}
\title{Extract plm model characteristics}
\usage{
extract_model_plm_stats(df_with_models)
}
\arguments{
\item{df_with_models}{}
}
\description{
Extract plm model characteristics
}

