% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_expression.R
\name{make_expression}
\alias{make_expression}
\title{Turn the right hand side variables into an expression.}
\usage{
make_expression(x)
}
\arguments{
\item{x}{}
}
\description{
Turn the right hand side variables into an expression.
}

