% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{make_predictions}
\alias{make_predictions}
\title{Title}
\usage{
make_predictions(train_start_date, train_end_date, test_start_date,
  test_end_date, df, expression, area_todo)
}
\arguments{
\item{area_todo}{}
}
\description{
Title
}

