% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_coef_by_mkt.R
\name{setup_coef_by_mkt}
\alias{setup_coef_by_mkt}
\title{Set up coefficients by market}
\usage{
setup_coef_by_mkt(df_with_models)
}
\arguments{
\item{df_with_models}{}
}
\description{
Set up coefficients by market
}

